Add new feature, show popup on double-click
Added a new feature where a dictionary popup appears when double-clicked on a text in a website. Three new files were created for this feature: popup_in_website.css for styling the popup, content-script.js for capturing double click events and handling the popup, and updated permissions in manifest.json to allow these actions on all URLS. The popup enables a user to quickly view the meaning of a word without leaving the webpage, thus improving user experience.